update: SimpleFactory ・コマンド出力
