+++ app/SimpleFactory/after/CheesePizza.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\SimpleFactory\After;
+
+class CheesePizza extends Pizza
+{
+    public function __construct()
+    {
+        $this->name  = 'チーズピザ';
+        $this->dough = 'レギュラー';
+        $this->sauce = 'マリアナソース';
+        $this->toppings[] = 'モッツァレラ';
+        $this->toppings[] = 'パルメザン';
+    }
+}
